Added V3 with update to CR123 Lithium Battery
Copy of the addon circuit that BMW used in the later boards and the "converted" boards with the little addon PCB. Uses a single CR123 3V lithium battery. This has a capacity of ~1600mAh compared to the combined 1900mAh of the original CR14250. The board uses Keystone 54 battery clips so the battery can be replaced without soldering.